[CHORE] 탈퇴, 퇴출 로직 도메인 변경
